Move schema to a separate file
