+++ emci/__main__.py
@@ -1,6 +1,7 @@
 from .rattler_build import build_with_rattler
 from .constants import RECIPES_SUBDIR_MAPPING, RECIPES_EMSCRIPTEN_DIR
 from .find_recipes_with_changes import find_recipes_with_changes
+from .schema import Recipe
 
 import sys
 import os
@@ -11,8 +12,6 @@ from pathlib import Path
 from typing import Optional
 import typer
 import yaml
-import re
-from pydantic import BaseModel, constr, field_validator
 
 app = typer.Typer(pretty_exceptions_enable=False)
 build_app = typer.Typer()
@@ -75,43 +74,6 @@ def bump_recipes_versions(target_branch_name: str):
     from .bot.bump_recipes_versions import bump_recipe_versions
 
     bump_recipe_versions(RECIPES_EMSCRIPTEN_DIR, target_branch_name)
-
-# --------------------------
-# Pydantic schema definitions
-# --------------------------
-
-class Source(BaseModel):
-    url: str
-    sha256: str
-
-    @field_validator("url")
-    @classmethod
-    def validate_url_or_template(cls, v: str) -> str:
-        # Allow templated strings like ${{ version }}
-        if "${{" in v and "}}" in v:
-            return v
-        # Otherwise, check it’s a valid URL
-        if not re.match(r"^https://.*\.(tar\.gz|zip)$", v):
-            raise ValueError("source.url must be a valid HTTPS link to .tar.gz or .zip, or contain templating")
-        return v
-
-    @field_validator("sha256")
-    @classmethod
-    def validate_sha256(cls, v: str) -> str:
-        if not re.match(r"^[0-9a-f]{64}$", v):
-            raise ValueError("source.sha256 must be a 64-character hex string")
-        return v
-
-class About(BaseModel):
-    license: str
-    license_file: str
-    license_family: Optional[str] = None
-
-
-class Recipe(BaseModel):
-    about: About
-    source: Source
-
 
 @build_app.command()
 def lint(old: str, new: str):
